Colorize tree charts based on minimum and maximum values
The colorization is done on the client side by navigating the tree and
mapping each of the values to a specific color. The mapping will respect
the order of a metric. E.g., for tests, large values are colored in
green while for complexities, large values are colored in red.