Add option to disable coverage column by default
This change adds global appearance configuration to
disable coverage column from showing by default.

The maven pipeline integration plugin now has a dependency
on this plugin which automatically adds the coverage column
which will always be "n/a" for projects that don't generate or
record coverage reports. This isn't a problem for custom
views, but the default "all" view includes all default columns
and can't be configured other wise. This change allows users
to remove the column from this view.